+++ client/CHANGELOG.md
@@ -4,9 +4,14 @@ All notable changes to the "vscode-abap-remote-fs" extension will be documented 
 
 Format based on [Keep a Changelog](http://keepachangelog.com/)
 
-## Unreleased
+## 0.5.23 2019-03-27
 
 ### Added
+
+- delete transport
+- reassign transport user
+- release transport
+- add user to transport
 
 ### Fixed